+++ History.md
@@ -1,6 +1,7 @@
 ### Unreleased
 
-* API change to remove Provision namespace
+* API change: remove Provision namespace
+* API change: remove nginx_directives option and add max_upload_mb
 
 ### Version 0.2.10
 2015-06-24